requirements.txt
bottle>=0.13.1
requests>=2.25.1
beaker>=1.11.0
bs4
pillow
pyjwt
bcrypt
gunicorn
cryptography
markdown